refactor: types and function names
