+++ package.json
@@ -53,7 +53,8 @@
   },
   "homepage": "https://github.com/MatteoGabriele/vue-gtag#readme",
   "peerDependencies": {
-    "vue": "^3.0.0"
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
@@ -66,7 +67,12 @@
     "vite": "^6.1.0",
     "vite-plugin-dts": "^4.5.0",
     "vitest": "^3.0.5",
-    "vue": "^3.0.0",
+    "vue": "^3.5.13",
     "vue-router": "^4.5.0"
+  },
+  "peerDependenciesMeta": {
+    "vue-router": {
+      "optional": true
+    }
   }
 }
