+++ src/api/ecommerce.ts
@@ -4,9 +4,6 @@ import type {
 } from "../types/gtag";
 import event from "./event";
 
-/**
- * @see https://developers.google.com/analytics/devguides/collection/ga4/reference/events?client_type=gtag#refund
- */
 export default function ecommerce(
   name: GtagEcommerceEventNames,
   params: GtagEcommerceParams,
