--- src/types/utils.ts
@@ -1,1 +1,0 @@
-export type Tail<T> = T extends [infer _, ...infer U] ? U : never;
